requirements.txt
Django
djangorestframework
django-otp
django-phonenumber-field
phonenumbers
twilio
requests
drf_yasg